Refactor handleInputTextChange to account for user deleting characters in input.
Since I'm keeping an internal map of the inputText which associates
each char with a uuid, I have to manipulate this data structure
along with the user's typing in the text input component. I check
the length of the text from the input component each time the
user interacts with it, and if it is shorter then the last time, I
just remove the last inputText object. If it is longer, then I can
create a new uuid and associate this with the char that was added.